Init format
